Drivers Updated

+++ RobotAxis/RobotAxis.cpp
@@ -1,7 +1,7 @@
 #include "RobotAxis.hpp"
 #include <iostream>
 
-RobotAxis::RobotAxis(Stepper &AXIS_MOTOR, const float STEP_ANGLE, const float AXIS_LENGTH) : _axisMotor(AXIS_MOTOR) {
+RobotAxis::RobotAxis(Stepper &AXIS_MOTOR, const double STEP_ANGLE, const double AXIS_LENGTH) : _axisMotor(AXIS_MOTOR) {
     _stepAngle = STEP_ANGLE;
     _axisLength = AXIS_LENGTH;
     _currentAngle = 0;
@@ -9,13 +9,12 @@ RobotAxis::RobotAxis(Stepper &AXIS_MOTOR, const float STEP_ANGLE, const float AX
     _axisMotor.setMaxSteps(360 / STEP_ANGLE);
 }
 
-void RobotAxis::rotate(float DESIRED_ANGLE) {
-    // return (_isPositive ? 1 : -1) * (static_cast<int>(DESIRED_ANGLE / _stepAngle));
+void RobotAxis::rotate(double DESIRED_ANGLE) {
 	_axisMotor.relStep((_isPositive ? 1 : -1) * (static_cast<int>(DESIRED_ANGLE / _stepAngle)));
     updatePosition();
 }
 
-void RobotAxis::rotate(float DESIRED_ANGLE, float OMEGA) {
+void RobotAxis::rotate(double DESIRED_ANGLE, double OMEGA) {
 	_axisMotor.velStep((_isPositive ? 1 : -1) * (static_cast<int>(DESIRED_ANGLE / _stepAngle)), OMEGA);
     updatePosition();
 }
@@ -24,11 +23,11 @@ void RobotAxis::updatePosition() {
     _currentAngle = (_isPositive ? 1.0f : -1.0f) * _axisMotor.getCurrentPosition() * _stepAngle;
 }
 
-void RobotAxis::goToAngle(float DESIRED_ANGLE) {
+void RobotAxis::goToAngle(double DESIRED_ANGLE) {
     rotate(DESIRED_ANGLE - _currentAngle);
 }
 
-void RobotAxis::goToAngle(float DESIRED_ANGLE, float OMEGA) {
+void RobotAxis::goToAngle(double DESIRED_ANGLE, double OMEGA) {
     rotate(DESIRED_ANGLE - _currentAngle, OMEGA);
 }
 
@@ -36,14 +35,14 @@ void RobotAxis::setDirection(bool IS_POSITIVE) {
     _isPositive = IS_POSITIVE;
 }
 
-void RobotAxis::setMaxVelocity(float MAX_VELOCITY) {
+void RobotAxis::setMaxVelocity(double MAX_VELOCITY) {
     _axisMotor.setMaxVelocity(MAX_VELOCITY);
 }
 
-void RobotAxis::setAcceleration(float ACCELERATION) {
+void RobotAxis::setAcceleration(double ACCELERATION) {
     _axisMotor.setAcceleration(ACCELERATION);
 }
 
-float RobotAxis::getCurrentAngle() {    return _currentAngle;   }
-float RobotAxis::getJointLength()  {    return _axisLength;     }
-float RobotAxis::getStepAngle()    {    return _stepAngle;      }
+double RobotAxis::getCurrentAngle() {    return _currentAngle;   }
+double RobotAxis::getJointLength()  {    return _axisLength;     }
+double RobotAxis::getStepAngle()    {    return _stepAngle;      }
